remove the need for an enterprise metabot provider and move the code into oss land, updated reference and remove some code no longer needed due to this change
