+++ enterprise/frontend/src/metabase-enterprise/metabot/tests/error-boundary.unit.spec.tsx
@@ -7,6 +7,7 @@ import { setupEnterprisePlugins } from "__support__/enterprise";
 import { mockSettings } from "__support__/settings";
 import { renderWithProviders, screen } from "__support__/ui";
 import { FIXED_METABOT_IDS } from "metabase/metabot/constants";
+import { MetabotProvider } from "metabase/metabot/context";
 import {
   createMockTokenFeatures,
   createMockUser,
@@ -14,7 +15,6 @@ import {
 import { createMockState } from "metabase-types/store/mocks";
 
 import { Metabot } from "../components/Metabot";
-import { MetabotProvider } from "../context";
 import { metabotReducer } from "../state";
 import { getMetabotInitialState } from "../state/reducer-utils";
