+++ frontend/src/metabase/plugins/oss/ai.ts
@@ -2,9 +2,8 @@ import type { BaseQueryFn } from "@reduxjs/toolkit/query";
 import type { TypedUseLazyQuery } from "@reduxjs/toolkit/src/query/react/buildHooks";
 import type { Extension } from "@uiw/react-codemirror";
 import type { ComponentType } from "react";
-import React from "react";
+import type React from "react";
 
-import type { MetabotContext } from "metabase/metabot";
 import { PluginPlaceholder } from "metabase/plugins/components/PluginPlaceholder";
 import type Question from "metabase-lib/v1/Question";
 import type {
@@ -61,9 +60,6 @@ type PluginMetabotType = {
     hide?: boolean;
     config?: PluginMetabotConfig;
   }) => React.ReactElement | null;
-  defaultMetabotContextValue: MetabotContext;
-  MetabotContext: React.Context<MetabotContext>;
-  getMetabotProvider: () => ComponentType<{ children: React.ReactNode }>;
   getAdminPaths: () => AdminPath[];
   getAdminRoutes: () => React.ReactElement;
   getMetabotRoutes: () => React.ReactElement | null;
@@ -97,17 +93,6 @@ type PluginMetabotType = {
   MetabotThinkingStyles: { [key: string]: string };
 };
 
-const getDefaultMetabotContextValue = (): MetabotContext => ({
-  prompt: "",
-  setPrompt: () => {},
-  promptInputRef: undefined,
-  getChatContext: () => ({}) as any,
-  registerChatContextProvider: () => () => {},
-});
-
-const defaultMetabotContextValue: MetabotContext =
-  getDefaultMetabotContextValue();
-
 const getDefaultPluginAiSqlFixer = (): PluginAiSqlFixer => ({
   FixSqlQueryButton: PluginPlaceholder,
 });
@@ -128,16 +113,6 @@ const getDefaultPluginMetabot = (): PluginMetabotType => ({
   isEnabled: () => false,
   Metabot: (_props: { hide?: boolean; config?: PluginMetabotConfig }) =>
     null as React.ReactElement | null,
-  defaultMetabotContextValue,
-  MetabotContext: React.createContext(defaultMetabotContextValue),
-  getMetabotProvider: () => {
-    return ({ children }) =>
-      React.createElement(
-        PLUGIN_METABOT.MetabotContext.Provider,
-        { value: PLUGIN_METABOT.defaultMetabotContextValue },
-        children,
-      );
-  },
   getAdminPaths: () => [],
   getAdminRoutes: () => PluginPlaceholder as unknown as React.ReactElement,
   getMetabotRoutes: () => null,
